requirements1.txt
numpy==1.24.4
Pillow==10.1.0
PyYAML==6.0.1
six==1.16.0
